requirements.txt
#list of python packages
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
oauth2client
typing-extensions
